Assignment 2 - Doubly Linked List
