+++ CLASS_01_PICS/Ajustamos a (20), valor esperado, remplazado por el (2) oldValue  06.png
