+++ CLASS_01_PICS/Detectamos, que lo que se espera, no es lo actual, ajustamos 05.png
