+++ CLASS_01_PICS/dotnet --version (en este caso .NET 8 )_ 00png.png
