+++ CLASS_01_PICS/dotnet build(Exito con .NET8 )_ 01.png
