+++ CLASS_01_PICS/dotnet test (NotImplementException - Procedemos investigar en la linea 53 de lnkList  )_ 04.png
