+++ CLASS_02_CODE/DoublyLnkList.cs
@@ -226,4 +226,6 @@ public class DoublyLnkList<T> : ILnkList<T>
         return false;
     }
 
-}
+    }
+
+
