+++ CLASS_02_CODE/DoublyLnkListTests.cs
@@ -2,7 +2,7 @@ namespace Tests;
 
 public class DoublyLnkListTests
 {
-
+   
     [Fact]
     // Es O(1) por que accede directamente al indice del elemento
     public void GetByIndex_UsingIndexer()
@@ -46,7 +46,7 @@ public class DoublyLnkListTests
         Assert.Throws<IndexOutOfRangeException>(
             () => ll.Get(1));
     }
-
+   
     [Fact]
     /* Es O(1) ya que especifica lo que busca sin importar el tamaño
     de la lista, este metodo busca en una lista vacia por lo tanto
@@ -54,7 +54,7 @@ public class DoublyLnkListTests
     hare un  throw new ArgumentOutOfRangeException por que no esta
     dentro del rango.
    */
-
+    
     public void GetByIndex_OutOfRange_Negative()
     {
         var ll = new DoublyLnkList<int>();
@@ -63,7 +63,7 @@ public class DoublyLnkListTests
             () => ll.Get(-1));
     }
 
-
+    
     [Fact]
     /* Es O(n) este metodo itera la lista hasta encontrar el elemento
      en el indice indicado, lo cual en el peor de los casos si el indice
@@ -79,10 +79,10 @@ public class DoublyLnkListTests
         Assert.Throws<IndexOutOfRangeException>(
             () => ll.Get(2));
     }
-
+  
     //************************************************************
 
-
+    
     //-**************************-> ANY<---****************************
     [Fact]
     /* Es O(n) este metodo itera una lista vacia donde se busca a el elemento "a"
@@ -96,8 +96,8 @@ public class DoublyLnkListTests
 
         Assert.False(ll.Any(value => value == 'a'));
     }
-
-
+    
+    
     /* Es O(1) en este ocacion tenemos una lista de characters, a,b,c you and me xD
      donde si encontramos el char 'b', retornamos true, es Big O(1) por que tenemos un
      apuntador directo 'b' y esta en la lista.
@@ -114,10 +114,10 @@ public class DoublyLnkListTests
 
         Assert.True(ll.Any(value => value == 'b'));
     }
-
+    
     //-*********************-> ANY<---****************************
 
-
+    
     //-*********************-> First<---****************************
     [Fact]
     // First(){}
@@ -126,7 +126,7 @@ public class DoublyLnkListTests
      * lo cual indica que si la lista es gigante, no importaria por que ya esta en
      * la secuencia inicial
      */
-
+    
     public void First_AfterPrepend()
     {
         var ll = new DoublyLnkList<int>();
@@ -151,25 +151,25 @@ public class DoublyLnkListTests
 
         Assert.Equal(30, ll.First());
     }
-
-
+    
+   
 
     [Fact]
     /* Es O(1) en este caso tenemos una lista vacia, y el Assert nos arroja 
      un Throws<InvalidOperationException si en ll.First()); no se encuentra 
      el primer elemento  First_Empty() xD 
    */
-
+    
     public void First_Empty()
     {
         var ll = new DoublyLnkList<int>();
 
         Assert.Throws<InvalidOperationException>(() => ll.First());
     }
-
+    
     //-*********************-> First<---****************************
 
-
+    
     //-*********************-> Count<---****************************
     [Fact]
     /* En este caso no es relevante en performance, ya que el metodo devuelve el conteo
@@ -200,7 +200,7 @@ public class DoublyLnkListTests
     }
 
     [Fact]
-
+    
     /* De igual forma, en este caso no es relevante a performance
      * ya que aunque se agregen 3 elementos  Assert.Equal espera 3
      * elementos y estos son los unicos agregados y por ello el count es 3
@@ -217,15 +217,15 @@ public class DoublyLnkListTests
 
         Assert.Equal(3, ll.Count());
     }
-    //-*********************-> Count<---****************************
+//-*********************-> Count<---****************************
 
 
-    //-*********************-> ToEnumerable<---****************************
+//-*********************-> ToEnumerable<---****************************
     [Fact]
     /* O(1) en este caso, ToEnumerable retorna vacio, por lo tanto 
     ll esta vacio   Assert.Equal se asegura de eso. 
     por lo tanto no itera en nada, ya que retorna que esta vacio el ToEnumerable
-    */
+    */ 
     public void ToEnumerable_Empty()
     {
         var ll = new DoublyLnkList<int>();
@@ -237,7 +237,7 @@ public class DoublyLnkListTests
     /* O(1) en este caso, ToEnumerable retorna vacio, por lo tanto
    ll esta vacio   Assert.Equal se asegura de eso.
    por lo tanto no itera en nada, ya que retorna que esta vacio el ToEnumerable
-   */
+   */ 
     public void ToEnumerable_OneElement()
     {
         var ll = new DoublyLnkList<int>();
@@ -254,7 +254,7 @@ public class DoublyLnkListTests
      * Y como de igual forma tiene que comparar el array debe de iterar
      * para ofrecer el resultado de la comparacion.
      */
-
+    
     public void ToEnumerable_ManyElement()
     {
         var ll = new DoublyLnkList<int>();
@@ -265,7 +265,7 @@ public class DoublyLnkListTests
 
         Assert.Equal(new[] { 1, 2, 3 }, ll.ToEnumerable());
     }
-
+    
     [Fact]
     /*  O(n) ya que itera los elementos de la lista
      emepezando desde el tail y nos dice cada valor en reversa 
@@ -299,9 +299,9 @@ public class DoublyLnkListTests
 
         Assert.Equal(new[] { 2, 1 }, ll.ToEnumerable());
     }
-    //-*********************-> ToEnumerable <---****************************
+//-*********************-> ToEnumerable <---****************************
 
-    //-*********************->RemoveAt <---****************************
+//-*********************->RemoveAt <---****************************
     [Fact]
     /*
      *O(1) como tiene un apuntador que es el indice 0
@@ -348,7 +348,7 @@ public class DoublyLnkListTests
 
         Assert.Equal(new[] { "a", "b" }, list.ToEnumerable());
     }
-
+    
     [Fact]
     /*
      * seria o(1) por que solo requiere actualizar las referencias
@@ -376,12 +376,12 @@ public class DoublyLnkListTests
 
         Assert.Equal(new[] { "a", "b", "c" }, list.ToEnumerable());
     }
-
-
-
+    
+    
+    
 
     [Fact]
-
+    
     /*
      *  Es O(1) ya que el indice es negativo, los cual esta fuera de rango,
      * entonces el metodo retorna sin tener que recorrer nada.
@@ -397,7 +397,7 @@ public class DoublyLnkListTests
     }
 
     //-*********************-> RemoveAt <---****************************
-
+    
     //-*********************-> Remove<---****************************
     [Fact]
     /*
@@ -440,7 +440,7 @@ public class DoublyLnkListTests
 
         Assert.Equal(new[] { "a", "b", "c" }, list.ToEnumerable());
     }
-
+    
     [Fact]
     /*
      *  O(n) recorre la lista completa, y n es largo de la misma,
@@ -455,7 +455,7 @@ public class DoublyLnkListTests
 
         Assert.Equal(new[] { "a", "b" }, list.ToEnumerable());
     }
-
+    
     [Fact]
     /* 
      *es O(1) ya que busca desde el head y es firts, 
@@ -468,6 +468,6 @@ public class DoublyLnkListTests
 
         Assert.Equal(new[] { "b", "c" }, list.ToEnumerable());
     }
-
+    
     //-*********************-> Remove <---****************************
 }
